chore(eth): refactor eth API module into separate pieces in new pkg
