+++ itests/eth_fee_history_test.go
@@ -16,7 +16,7 @@ import (
 	"github.com/filecoin-project/lotus/chain/types/ethtypes"
 	"github.com/filecoin-project/lotus/itests/kit"
 	"github.com/filecoin-project/lotus/lib/result"
-	"github.com/filecoin-project/lotus/node/impl/full"
+	"github.com/filecoin-project/lotus/node/impl/gasutils"
 )
 
 // calculateExpectations calculates the expected number of items to be included in the response
@@ -171,7 +171,7 @@ func TestEthFeeHistory(t *testing.T) {
 	for _, arr := range *history.Reward {
 		require.Equal(3, len(arr))
 		for _, item := range arr {
-			require.Equal(ethtypes.EthBigInt(types.NewInt(full.MinGasPremium)), item)
+			require.Equal(ethtypes.EthBigInt(types.NewInt(gasutils.MinGasPremium)), item)
 		}
 	}
